[feat] add loading spinner
