+++ teamplan/Sources/Views/Loading/LoadingView.swift
@@ -1,41 +1,33 @@
 //
-//  LoadingView2.swift
+//  LoadingView.swift
 //  teamplan
 //
-//  Created by sungyeon on 2023/11/06.
+//  Created by sungyeon on 2023/11/08.
 //  Copyright © 2023 team1os. All rights reserved.
 //
 
 import SwiftUI
 
-struct LoadingView<Content>: View where Content: View {
-
-    @Binding var isShowing: Bool
-    var content: () -> Content
-
+struct LoadingView: View {
     var body: some View {
-        GeometryReader { geometry in
-            ZStack(alignment: .center) {
-
-                self.content()
-                    .disabled(self.isShowing)
-                    .blur(radius: self.isShowing ? 3 : 0)
-
-                ActivityIndicator(isAnimating: .constant(true), style: .large)
-                    .opacity(self.isShowing ? 1 : 0)
-//                VStack {
-//                    Text("Loading...")
-//
-//                }
-//                .frame(width: geometry.size.width / 2,
-//                       height: geometry.size.height / 5)
-//                .background(Color.secondary.colorInvert())
-//                .foregroundColor(Color.primary)
-//                .cornerRadius(20)
-                
-
-            }
+        ZStack {
+            Color(.black)
+                .ignoresSafeArea()
+                .opacity(0.2)
+                .blur(radius: 3)
+            
+            ProgressView()
+                .progressViewStyle(
+                    CircularProgressViewStyle(tint: .gray)
+                )
+                .scaleEffect(3)
         }
+        
     }
+}
 
+struct LoadingView_Previews: PreviewProvider {
+    static var previews: some View {
+        LoadingView()
+    }
 }
